fix bug

+++ data/contrastive_dataset.py
@@ -149,7 +149,7 @@ class ContrastiveResponseSelectionDataset(Dataset):
         current_feature["res_sel"]["eot_pos"] = torch.tensor(eot_pos).long()
         current_feature["res_sel"]["label"] = torch.tensor(example.label).float()
         if hasattr(example, 'soft_logits'):
-            current_feature["res_sel"]["soft_logits"] = torch.tensor(example.soft_logits)
+            current_feature["res_sel"]["soft_logits"] = torch.tensor(example.soft_logits).float()
 
         # when the response is the ground truth, append it to utterances.
         if int(example.label) == 1:
